+++ src/assault/game/loading/ResourcePreloader.java
@@ -41,7 +41,7 @@ public class ResourcePreloader extends Paintable implements Disposable{
 	private SAXBuilder builder;
 	private boolean started = false;
 
-	public ResourcePreloader(String[] mods, double w, double h) {
+	public ResourcePreloader(String[] mods, float w, float h) {
 		super(0, 0, w, h);
 		modNames = mods;
 	}
